+++ src/ts/sv/process.ts
@@ -59,7 +59,10 @@ export const processSpinVditorSVDOM = (html: string, vditor: IVditor) => {
 export const processPreviousMarkers = (spanElement: HTMLElement) => {
     const spanType = spanElement.getAttribute("data-type");
     let previousElement = spanElement.previousElementSibling;
-    let markerText = (spanType && spanType !== "text") ? spanElement.textContent : ""; // 有内容的子列表，在其 marker 后换行
+    // 有内容的子列表/标题，在其 marker 后换行
+    let markerText = (spanType && spanType !== "text" && spanType !== "heading-marker" &&
+        spanType !== "yaml-front-matter-open-marker" && spanType !== "yaml-front-matter-close-marker") ?
+        spanElement.textContent : "";
     let hasNL = false;
     while (previousElement && !hasNL) {
         const previousType = previousElement.getAttribute("data-type");